+++ library/src/auxiliary/rocauxiliary_larfg.hpp
@@ -39,8 +39,9 @@
 ROCSOLVER_BEGIN_NAMESPACE
 
 template <typename T, std::enable_if_t<!rocblas_is_complex<T>, int> = 0>
-__device__ void run_set_taubeta(T* tau, T* norms, T* alpha)
+__device__ void run_set_taubeta(T* tau, T* norms, T* alpha, T* beta)
 {
+    const auto ignore_beta = (beta == nullptr);
     if(norms[0] > 0)
     {
         T n = sqrt(norms[0] + alpha[0] * alpha[0]);
@@ -53,17 +54,32 @@ __device__ void run_set_taubeta(T* tau, T* norms, T* alpha)
         tau[0] = (n - alpha[0]) / n;
 
         // beta:
-        alpha[0] = n;
+        if(ignore_beta)
+        {
+            alpha[0] = n;
+        }
+        else
+        {
+            beta[0] = n;
+            alpha[0] = 1;
+        }
     }
     else
     {
         norms[0] = 1;
         tau[0] = 0;
+
+        // beta:
+        if(!ignore_beta)
+        {
+            beta[0] = alpha[0];
+            alpha[0] = 1;
+        }
     }
 }
 
 template <typename T, std::enable_if_t<rocblas_is_complex<T>, int> = 0>
-__device__ void run_set_taubeta(T* tau, T* norms, T* alpha)
+__device__ void run_set_taubeta(T* tau, T* norms, T* alpha, T* beta)
 {
     using S = decltype(std::real(T{}));
     S r, rr, ri, ar, ai;
@@ -72,6 +88,7 @@ __device__ void run_set_taubeta(T* tau, T* norms, T* alpha)
     ai = alpha[0].imag();
     S m = ai * ai;
 
+    const auto ignore_beta = (beta == nullptr);
     if(norms[0].real() > 0 || m > 0)
     {
         m += ar * ar;
@@ -92,30 +109,49 @@ __device__ void run_set_taubeta(T* tau, T* norms, T* alpha)
         tau[0] = rocblas_complex_num<S>(rr, ri);
 
         // beta:
-        alpha[0] = n;
+        if(ignore_beta)
+        {
+            alpha[0] = n;
+        }
+        else
+        {
+            beta[0] = n;
+            alpha[0] = 1;
+        }
     }
     else
     {
         norms[0] = 1;
         tau[0] = 0;
+
+        // beta:
+        if(!ignore_beta)
+        {
+            beta[0] = alpha[0];
+            alpha[0] = 1;
+        }
     }
 }
 
-template <typename T, typename I, typename U>
+template <typename T, typename I, typename U, typename UB>
 ROCSOLVER_KERNEL void set_taubeta(T* tauA,
                                   const rocblas_stride strideP,
                                   T* norms,
                                   U alphaA,
                                   const rocblas_stride shiftA,
-                                  const rocblas_stride strideA)
+                                  const rocblas_stride strideA,
+                                  UB betaA,
+                                  const rocblas_stride shiftb,
+                                  const rocblas_stride strideb)
 {
     I bid = hipBlockIdx_x;
 
     // select batch instance
     T* alpha = load_ptr_batch<T>(alphaA, bid, shiftA, strideA);
+    T* beta = betaA ? load_ptr_batch<T>(betaA, bid, shiftb, strideb) : nullptr;
     T* tau = tauA + bid * strideP;
 
-    run_set_taubeta<T>(tau, norms + bid, alpha);
+    run_set_taubeta<T>(tau, norms + bid, alpha, beta);
 }
 
 template <typename T, typename I>
@@ -183,11 +219,14 @@ rocblas_status
     return rocblas_status_continue;
 }
 
-template <typename T, typename I, typename U, bool COMPLEX = rocblas_is_complex<T>>
+template <typename T, typename I, typename U, typename UB, bool COMPLEX = rocblas_is_complex<T>>
 rocblas_status rocsolver_larfg_template(rocblas_handle handle,
                                         const I n,
                                         U alpha,
                                         const rocblas_stride shifta,
+                                        UB beta,
+                                        const rocblas_stride shiftb,
+                                        const rocblas_stride strideb,
                                         U x,
                                         const rocblas_stride shiftx,
                                         const I incx,
@@ -211,11 +250,17 @@ rocblas_status rocsolver_larfg_template(rocblas_handle handle,
 
     // if n==1 return tau=0
     dim3 gridReset(1, batch_count, 1);
+    dim3 setDiag(batch_count, 1, 1);
     dim3 threads(1, 1, 1);
     if(n == 1 && !COMPLEX)
     {
         ROCSOLVER_LAUNCH_KERNEL(reset_batch_info<T>, gridReset, threads, 0, stream, tau, strideP, 1,
                                 0);
+        if(beta != nullptr)
+        {
+            ROCSOLVER_LAUNCH_KERNEL((set_diag<T>), setDiag, threads, 0, stream, beta, shiftb,
+                                    strideb, alpha, shifta, n, stridex, (I)1, true);
+        }
         return rocblas_status_success;
     }
 
@@ -229,8 +274,8 @@ rocblas_status rocsolver_larfg_template(rocblas_handle handle,
         HIP_CHECK(hipGetDeviceProperties(&deviceProperties, device));
         if(deviceProperties.warpSize >= 64)
         {
-            return larfg_run_small(handle, n, alpha, shifta, stridex, x, shiftx, incx, stridex, tau,
-                                   strideP, batch_count);
+            return larfg_run_small(handle, n, alpha, shifta, stridex, beta, shiftb, strideb, x,
+                                   shiftx, incx, stridex, tau, strideP, batch_count);
         }
     }
 
@@ -246,13 +291,32 @@ rocblas_status rocsolver_larfg_template(rocblas_handle handle,
     // set value of tau and beta and scalling factor for vector x
     // alpha <- beta, norms <- scaling
     ROCSOLVER_LAUNCH_KERNEL((set_taubeta<T, I>), dim3(batch_count), dim3(1), 0, stream, tau,
-                            strideP, norms, alpha, shifta, stridex);
+                            strideP, norms, alpha, shifta, stridex, beta, shiftb, strideb);
 
     // compute vector v=x*norms
     rocblasCall_scal<T>(handle, n - 1, norms, 1, x, shiftx, incx, stridex, batch_count);
 
     rocblas_set_pointer_mode(handle, old_mode);
     return rocblas_status_success;
+}
+
+template <typename T, typename I, typename U, bool COMPLEX = rocblas_is_complex<T>>
+rocblas_status rocsolver_larfg_template(rocblas_handle handle,
+                                        const I n,
+                                        U alpha,
+                                        const rocblas_stride shifta,
+                                        U x,
+                                        const rocblas_stride shiftx,
+                                        const I incx,
+                                        const rocblas_stride stridex,
+                                        T* tau,
+                                        const rocblas_stride strideP,
+                                        const I batch_count,
+                                        T* work,
+                                        T* norms)
+{
+    return rocsolver_larfg_template(handle, n, alpha, shifta, (T*)nullptr, 0, 0, x, shiftx, incx,
+                                    stridex, tau, strideP, batch_count, work, norms);
 }
 
 ROCSOLVER_END_NAMESPACE