+++ library/src/include/rocsolver_run_specialized_kernels.hpp
@@ -59,12 +59,15 @@ rocblas_status larf_run_small(rocblas_handle handle,
                               const I batch_count);
 
 // larfg
-template <typename T, typename I, typename U>
+template <typename T, typename I, typename U, typename UB>
 rocblas_status larfg_run_small(rocblas_handle handle,
                                const I n,
                                U alpha,
                                const rocblas_stride shiftA,
                                const rocblas_stride strideA,
+                               UB beta,
+                               const rocblas_stride shiftB,
+                               const rocblas_stride strideB,
                                U x,
                                const rocblas_stride shiftX,
                                const I incX,